sdk-core/subgraph cleanup pt1
## package.json
- remove sdk-core from root package.json tests

## ci.feature
- remove eager testing in ci.feature

## ci.canary
- remove eager testing in ci.canary
- call.setup-deploy-and-test-local-subgraph.yml only tests local subgraphs, even schema generation
- consolidate `runQueryTests.ts` and `queryTests.ts`
