+++ package.json
@@ -16,7 +16,7 @@
         "build-essentials": "set -e;for i in ethereum-contracts js-sdk sdk-core sdk-redux;do yarn workspace @superfluid-finance/$i build;done",
         "build-for-contracts-dev": "set -e;for i in ethereum-contracts js-sdk;do yarn workspace @superfluid-finance/$i build;done",
         "clean": "rm -rf node_modules; rm -rf packages/*/node_modules; yarn workspace @superfluid-finance/ethereum-contracts clean",
-        "test": "set -e;for i in ethereum-contracts sdk-core;do yarn workspace @superfluid-finance/$i test;done",
+        "test": "set -e;for i in ethereum-contracts;do yarn workspace @superfluid-finance/$i test;done",
         "manage-versions": "lerna version --exact --no-git-tag-version --preid rc",
         "check-updates": "ncu --target minor --dep dev",
         "show-versions": "lerna ls --long",
