+++ packages/sdk-core/test/5_subgraph.test.ts
@@ -6,14 +6,18 @@ import {
     testListenerInitialization,
     testQueryClassFunctions,
 } from "../previous-versions-testing/queryTests";
-import { getSubgraphEndpoint } from "../previous-versions-testing/runQueryTests";
+import { getSubgraphEndpoint } from "../previous-versions-testing/queryTests";
 import { expect } from "chai";
 
 describe("Subgraph Tests", () => {
     let query: Query;
 
     const chainIdToUse = getChainId();
-    const customSubgraphQueriesEndpoint = getSubgraphEndpoint(chainIdToUse);
+
+    // we will either test a singular subgraph_endpoint
+    // or we test all of the subgraph endpoints
+    const customSubgraphQueriesEndpoint =
+        process.env.SUBGRAPH_ENDPOINT || getSubgraphEndpoint(chainIdToUse);
 
     before(() => {
         query = new Query({
@@ -38,31 +42,28 @@ describe("Subgraph Tests", () => {
             await testListenerInitialization(query);
         });
 
-        it("Should have the correct subgraph endpoints", async () => {
+        it("Should have same schema entity as deployed subgraph endpoints", async () => {
             const resolverDataArray = Array.from(
                 chainIdToResolverDataMap.values()
             ) as NetworkData[];
-            await Promise.all(
-                resolverDataArray.map(async (x) => {
-                    // @note if SUBGRAPH_RELEASE_TAG is feature, we only test goerli
-                    const isValidFeatureEndpoint =
-                        process.env.SUBGRAPH_RELEASE_TAG === "feature" &&
-                        x.networkName === "eth-goerli";
-                    // @note else we test all networks
-                    const isNotFeatureEndpoint =
-                        process.env.SUBGRAPH_RELEASE_TAG !== "feature";
+            if (process.env.SUBGRAPH_ENDPOINT === "") {
+                await Promise.all(
+                    resolverDataArray.map(async (x) => {
+                        const isNotFeatureEndpoint =
+                            process.env.SUBGRAPH_RELEASE_TAG !== "feature";
 
-                    // @note MUST be eth-goerli for feature release tag OR NOT feature release tag
-                    if (isValidFeatureEndpoint || isNotFeatureEndpoint) {
-                        const query = new Query({
-                            customSubgraphQueriesEndpoint:
-                                x.subgraphAPIEndpoint,
-                        });
-                        const event = await query.listEvents({}, { take: 1 });
-                        expect(event.data.length).to.be.greaterThan(0);
-                    }
-                })
-            );
+                        if (isNotFeatureEndpoint) {
+                            const query = new Query({
+                                customSubgraphQueriesEndpoint:
+                                    x.subgraphAPIEndpoint,
+                            });
+                            await testGetAllEventsQuery(query);
+                        }
+                    })
+                );
+            } else {
+                testGetAllEventsQuery(query);
+            }
         });
     });
 });
